#1639 Inba Releases page redesign
